Extend data-type to initial-config-primitive
The config-primitive defines a data-type, which is an enumerator of
STRING, INTEGER, and BOOLEAN, that informs consumers of the VNFD of what
type to treat the value as.

Yang itself is loosely typed, treating everything as a string, but
specifying the data-type is required when integrating with
strongly-typed languages. We've experienced this issue, documented in
bug #502, where the service primitives defined via the VCA are
strongly-typed and will fail if used with a primitive expecting a
non-STRING value.

This extends the data-type key to initial-config-primitive, allowing the
operator to define what type of data is expected. This will then be
consumed by N2VC to properly convert the value appropriately.

Signed-off-by: Adam Israel <adam.israel@canonical.com>